Change revert proxy handler
