+++ bUtility.ReverseProxy/ProxyHandler.cs
@@ -19,52 +19,33 @@ using System.Xml;
 namespace bUtility.ReverseProxy
 {
     public class ReverseProxyHandler : DelegatingHandler
-    {
-        public class UtilityClient : HttpClient
-        {
-            public UtilityClient()
-                : base()
-            {
-            }
-            public UtilityClient(HttpMessageHandler messageHandler)
-                : base(messageHandler)
-            {
-
-            }
-            public System.Threading.Tasks.Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
-                CancellationToken cancellationToken, string authHeaderSchema, string authToken)
-            {
-                var httpFriendlyToken = Convert.ToBase64String(Encoding.UTF8.GetBytes(authToken));
-                request.Headers.Authorization = new AuthenticationHeaderValue(authHeaderSchema, httpFriendlyToken);
-                return base.SendAsync(request, cancellationToken);
-            }
-        }
-
+    { 
         string WebApiSourceUrl { get; set; }
         string WebApiDestinationUrl { get; set; }
-        Action<UtilityClient> PrepareApiCall { get; set; }
+        Action<HttpRequestMessage> PrepareApiCall { get; set; }
         Action<HttpResponseMessage> PrepareResponse { get; set; }
+        private readonly Func<HttpClient> HttpClientProvider;
         readonly ILogger Logger;
-        public ReverseProxyHandler(string webApiSourceUrl, string webApiDestinationUrl,
-            Action<UtilityClient> prepareApiCall, Action<HttpResponseMessage> prepareResponse,
-            ILogger logger)
+        public ReverseProxyHandler(string webApiSourceUrl, string webApiDestinationUrl, 
+            Func<HttpClient> httpClientProvider, Action<HttpRequestMessage> prepareApiCall,
+            Action<HttpResponseMessage> prepareResponse, ILogger logger)
         {
             if (string.IsNullOrEmpty(webApiSourceUrl))
                 throw new ArgumentNullException("webApiSourceUrl");
             if (string.IsNullOrEmpty(webApiDestinationUrl))
                 throw new ArgumentNullException("webApiDestinationUrl");
+            if (httpClientProvider == null)
+                throw new ArgumentNullException(nameof(httpClientProvider));
             WebApiSourceUrl = webApiSourceUrl.ToLowerInvariant();
             WebApiDestinationUrl = webApiDestinationUrl.ToLowerInvariant();
+            HttpClientProvider = httpClientProvider;
             PrepareApiCall = prepareApiCall;
             PrepareResponse = prepareResponse;
             Logger = logger;
         }
-        bool RemoteCertificateValidationCallback(Object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors)
-        {
-            return true;
-        }
 
-        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
+            CancellationToken cancellationToken)
         {
             Logger.Warn($"requestUri: {request.RequestUri}, method: {request.Method}");
             if (request.RequestUri.ToString().StartsWith(WebApiSourceUrl, StringComparison.InvariantCultureIgnoreCase))
@@ -72,11 +53,11 @@ namespace bUtility.ReverseProxy
                 try
                 {
                     UriBuilder forwardUri = new UriBuilder(request.RequestUri);
-                    var targetUri = new Uri(forwardUri.Uri.ToString().ToLowerInvariant().Replace(WebApiSourceUrl, WebApiDestinationUrl));
+                    var targetUri = new Uri(forwardUri.Uri.ToString().ToLowerInvariant()
+                        .Replace(WebApiSourceUrl, WebApiDestinationUrl));
                     Logger.Warn($"targetUri: {targetUri}");
 
-                    //send it on to the requested URL
-
+                    //send it on to the requested URL 
                     var apiRequest = new HttpRequestMessage(request.Method, targetUri);
 
                     apiRequest.Version = request.Version;
@@ -88,38 +69,34 @@ namespace bUtility.ReverseProxy
                         apiRequest.Content.Headers.ContentType = request.Content.Headers.ContentType;
                     }
 
-                    using (WebRequestHandler messageHandler = new WebRequestHandler())
+                    PrepareApiCall(apiRequest);
+                    HttpClient client = HttpClientProvider(); 
+
+                    var apiResponse = await client.SendAsync(apiRequest, HttpCompletionOption.ResponseHeadersRead,
+                        cancellationToken);
+                    var respData = await apiResponse.Content.ReadAsByteArrayAsync();
+
+                    var fResponse = request.CreateResponse();
+                    fResponse.Content = new ByteArrayContent(respData);
+                    fResponse.Content.Headers.ContentType = apiResponse.Content.Headers.ContentType;
+                    fResponse.Content.Headers.ContentDisposition = apiResponse.Content.Headers.ContentDisposition;
+                    fResponse.ReasonPhrase = apiResponse.ReasonPhrase;
+                    fResponse.StatusCode = apiResponse.StatusCode;
+                    fResponse.Version = apiResponse.Version;
+
+                    //populate fResponse Headers based on apiResponse Headers
+
+                    if (fResponse.StatusCode == HttpStatusCode.Unauthorized)
                     {
-                        messageHandler.ServerCertificateValidationCallback += RemoteCertificateValidationCallback;
-                        UtilityClient client = new UtilityClient(messageHandler);
-
-                        PrepareApiCall?.Invoke(client);
-                        //populate UtilityClient Headers based on claims, or other headers
-
-
-                        var apiResponse = await client.SendAsync(apiRequest, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
-                        var respData = await apiResponse.Content.ReadAsByteArrayAsync();
-
-                        var fResponse = request.CreateResponse();
-                        fResponse.Content = new ByteArrayContent(respData);
-                        fResponse.Content.Headers.ContentType = apiResponse.Content.Headers.ContentType;
-                        fResponse.Content.Headers.ContentDisposition = apiResponse.Content.Headers.ContentDisposition;
-                        fResponse.ReasonPhrase = apiResponse.ReasonPhrase;
-                        fResponse.StatusCode = apiResponse.StatusCode;
-                        fResponse.Version = apiResponse.Version;
-
-                        //populate fResponse Headers based on apiResponse Headers
-
-                        if (fResponse.StatusCode == HttpStatusCode.Unauthorized)
-                        {
-                            //we have a 401 in api. Clear the local (in memory AND sql security token cache) 
-                            //and send the 401 to the browser with clear FedAuth cookie. 
-                            //The browser will refresh and will call (another perhaps) web server with no cookies, 
-                            //but for html wif will redirect the browser to sts
-                            System.IdentityModel.Services.FederatedAuthentication.SessionAuthenticationModule.DeleteSessionTokenCookie();
-                        }
-                        return fResponse;
+                        //we have a 401 in api. Clear the local (in memory AND sql security token cache) 
+                        //and send the 401 to the browser with clear FedAuth cookie. 
+                        //The browser will refresh and will call (another perhaps) web server with no cookies, 
+                        //but for html wif will redirect the browser to sts
+                        System.IdentityModel.Services.FederatedAuthentication.SessionAuthenticationModule
+                            .DeleteSessionTokenCookie();
                     }
+
+                    return fResponse;
                 }
                 catch (Exception ex)
                 {
