+++ bUtility.ReverseProxy/bUtility.ReverseProxy.csproj
@@ -60,6 +60,7 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProxyHandler.cs" />
+    <Compile Include="UtilityClient.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
